+++ src/utils/PathwayActions.ts
@@ -16,7 +16,7 @@ import GridOptionsManager from '../managers/GridOptionsManager'
 import { ILayoutProperties } from '../modals/LayoutProperties'
 import { EGridType } from '../modals/GridSettings'
 import ConfirmationModal from '../modals/ConfirmationModal'
-import NGCHM from './NGCHM'
+//import NGCHM from './NGCHM'
 
 export default class PathwayActions {
   @observable
@@ -36,8 +36,7 @@ export default class PathwayActions {
   isCollaborative: boolean
   viewOperationsManager: ViewOperationsManager
   overlayUploader: any
-  isInIframe: boolean
-  ngchm: any
+  //isInIframe: boolean
 
   @observable
   enabledType: EGridType
@@ -49,7 +48,7 @@ export default class PathwayActions {
     handleOpen: (modalId: EModalType) => void,
     isCBioPortal: boolean,
     isCollaborative: boolean,
-    isInIframe: boolean
+    //isInIframe: boolean
   ) {
     this.pathwayHandler = pathwayHandler
     this.profiles = profiles
@@ -58,10 +57,6 @@ export default class PathwayActions {
     this.isCBioPortal = isCBioPortal
     this.isCollaborative = isCollaborative
     this.enabledType = EGridType.NONE
-    this.isInIframe = isInIframe
-    if (this.isInIframe) {
-      this.ngchm = new NGCHM(this.profiles, this)
-    }
   }
 
   emphasiseQueryGenes(queryGenes: string[]) {
@@ -271,12 +266,6 @@ export default class PathwayActions {
   @autobind
   highlightSelected() {
     this.editor.highlightSelected()
-    if (this.isInIframe) {
-      if (typeof this.ngchm === 'undefined') {
-        this.ngchm = new NGCHM(this.profiles,this)
-      }
-      this.ngchm.highlightSelected()
-    }
   }
 
   @autobind
@@ -365,12 +354,6 @@ export default class PathwayActions {
   @autobind
   removeAllHighlight() {
     this.editor.removeAllHighlight()
-    if (this.isInIframe) {
-      if (typeof this.ngchm === 'undefined') {
-        this.ngchm = new NGCHM(this.profiles,this)
-      }
-      this.ngchm.highlightSelected()
-    }
   }
 
   @autobind
@@ -432,10 +415,6 @@ export default class PathwayActions {
     this.undoRedoManager = undoRedoManager
     this.viewOperationsManager = viewOperationsManager
     this.gridOptionsManager = gridOptionsManager
-    if (typeof this.ngchm != 'undefined') {
-      // define editor for ngchm
-      this.ngchm.editorHandler(editor)
-    }
   }
 
   @autobind
